Move dogpile region code to separate file
This is necessary as so many places are now using dogpile region, and this is the way to avoid circular imports.
